drivers/intel/fsp2_0/notify: Fix compilation under x86_64
Reviewed-on: https://review.coreboot.org/c/coreboot/+/48168
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/drivers/intel/fsp2_0/notify.c
@@ -57,7 +57,7 @@ static void fsp_notify(enum fsp_notify_phase phase)
 
 static void fsp_notify_dummy(void *arg)
 {
-	enum fsp_notify_phase phase = (uint32_t)arg;
+	enum fsp_notify_phase phase = (uint32_t)(uintptr_t)arg;
 
 	display_mtrrs();
